requirements.txt
beautifulsoup4==4.4.1
boto==2.38.0
cffi==1.3.1
cryptography==1.1.1
Django==1.8.6
django-hvad==1.4.0
django-libs==1.66.12
django-paypal==0.2.7
django-paypal-express-checkout==1.8
django-require==1.0.8
django-require-s3==1.0.0
django-storages==1.1.8
djangorestframework==3.3.1
djangorestframework-jwt==1.7.2
enum34==1.1.1
factory-boy==2.6.0
fake-factory==0.5.3
idna==2.0
ipaddress==1.0.15
oauthlib==1.0.3
paypalrestsdk==1.11.4
Pillow==3.0.0
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
PyJWT==1.4.0
pyOpenSSL==0.15.1
python-openid==2.2.5
python-social-auth==0.2.13
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
sorl-thumbnail==12.3
stripe==1.27.1
wheel==0.24.0
